fix(api): parse memory_limit as float to avoid 32-bit overflow
og.php scaled memory_limit to bytes with an int accumulator. On a 32-bit
PHP build an int caps near 2.1e9, so a 2G-or-larger limit overflowed —
wrapping negative and tripping the insufficient-memory guard on a host that
actually has plenty, refusing to render the OG image. Parse as float, which
represents these byte counts exactly well past any real memory_limit.

+++ api/og.php
@@ -330,7 +330,11 @@ if (!function_exists('imagecreatetruecolor')) {
 
 $limit = ini_get('memory_limit');
 if ($limit !== false && $limit !== '' && $limit !== '-1') {
-    $val = (int) $limit;
+    // Parse as float, not int: on a 32-bit PHP build an int caps near 2.1e9, so a
+    // "2G"+ limit would overflow when scaled to bytes below (wrapping negative and
+    // tripping the "insufficient memory" guard on a host that has plenty). Floats
+    // represent these byte counts exactly well past any real memory_limit.
+    $val = (float) $limit;
     $last = strtolower(substr(trim($limit), -1));
     if ($last === 'g') {
         $val *= 1024 * 1024 * 1024;
